Enhanced Bend Rendering by utilizing available space, detect bend type as they exist in Guitar Pro for further reextensions

+++ Source/AlphaTab/Rendering/Utils/BeamingHelper.cs
@@ -155,9 +155,9 @@ namespace AlphaTab.Rendering.Utils
             {
                 if (Direction == BeamDirection.Up)
                 {
-                    return _beatLineXPositions[beat.Index].Up;
+                    return (int)_beatLineXPositions[beat.Index].Up;
                 }
-                return _beatLineXPositions[beat.Index].Down;
+                return (int)_beatLineXPositions[beat.Index].Down;
             }
             return 0;
         }
